[#21] RequestViewModel 설계 완료
